requirements.txt
flask
requests
passlib
mysql-connector
json